Added "New Event Type" button on Team settings
- added "New Event Type" button
- bugfix: changed back button behavior in Shell to have a specified back path
